Create index.js
As stated in the READ.ME, this Learning-React repository is for the code and learning aids I will be creating for React.

As I go through Modern React with Redux on Udemy, I am adding comments as notes to the code provided in the course. If in the future I forget anything, these comments will serve as a quick guide to steer me back on course.